pyproject.toml
[project]
name = "yaml-playwright"
version = "0.1.0"
description = "YAML-driven web automation framework powered by Playwright"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["playwright", "automation", "yaml", "web", "testing"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "playwright>=1.40.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jinja2>=3.0",
    "rich>=13.0",
    "click>=8.0",
    "apscheduler>=3.10",
    "python-dateutil>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pre-commit>=3.0",
    "ruff>=0.1",
    "mypy>=1.7",
]

[project.scripts]
yamlplaywright = "src.cli:main"
yamla = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = false
warn_return_any = true
